Separate site naming resources
